добавит наследование простым компонентам
