+++ src/view/trip-sort-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 const createTripSortTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <div class="trip-sort__item  trip-sort__item--day">
@@ -24,21 +24,8 @@ const createTripSortTemplate = () => (
     </form>`
 );
 
-export default class TripSortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
+export default class TripSortView extends AbstractView {
   get template() {
     return createTripSortTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
